fix: address follow-up portal item OpenAPI review feedback
